feat: alter app to avoid upload from not logged users and show the public stuff for non logged in

+++ packages/backend/src/controllers/files-controller.ts
@@ -3,6 +3,11 @@ import { fileRepository } from "../repository/sqlite";
 import { authMiddleware } from "../middleware";
 import { FilesService } from "../services/files-service";
 import { NotFoundError } from "../errors/not-found-error";
+import {
+  validateUserCanUploadPublicFiles,
+  validateUserCanWrite,
+} from "../utils/user-rights-validator";
+import { logger } from "../utils/logger-utils";
 
 export const filesController = (expressServer: Express) => {
   const filesService = new FilesService(fileRepository);
@@ -21,12 +26,23 @@ export const filesController = (expressServer: Express) => {
         })),
       );
     } catch (error) {
-      console.error("Error listing file:", error);
+      logger.error("Error listing file:", error);
       res.status(500).json({ error: "Internal server error" });
     }
   });
 
   expressServer.put("/api/files/:id", authMiddleware, async (req, res) => {
+    if (!validateUserCanWrite((req as any).userRights)) {
+      return res.status(403).json({ error: "Forbidden" });
+    }
+
+    if (
+      req.body.privacy &&
+      !validateUserCanUploadPublicFiles((req as any).userRights)
+    ) {
+      return res.status(403).json({ error: "Forbidden to set public privacy" });
+    }
+
     try {
       const { id } = req.params;
       const { parent, originalName } = req.body;
@@ -36,6 +52,7 @@ export const filesController = (expressServer: Express) => {
         parent,
         originalName,
         userId: (req as any).userId,
+        privacy: req.body.privacy,
       });
 
       res.json(updatedFile);
@@ -44,12 +61,16 @@ export const filesController = (expressServer: Express) => {
         return res.status(404).json({ error: error.message });
       }
 
-      console.error("Error updating file:", error);
+      logger.error("Error updating file:", error);
       return res.status(500).json({ error: "Internal server error" });
     }
   });
 
   expressServer.delete("/api/files/:id", authMiddleware, async (req, res) => {
+    if (!validateUserCanWrite((req as any).userRights)) {
+      return res.status(403).json({ error: "Forbidden" });
+    }
+
     try {
       const result = await filesService.delete({
         id: req.params.id,
@@ -61,7 +82,7 @@ export const filesController = (expressServer: Express) => {
       if (error instanceof NotFoundError) {
         return res.status(404).json({ error: error.message });
       }
-      console.error("Error deleting file:", error);
+      logger.error("Error deleting file:", error);
       return res.status(500).json({ error: "Internal server error" });
     }
   });
